+++ packages/app/imports/api/champions/champion-matchups/_collection.js
@@ -1,3 +1,0 @@
-import TimestampedCollection from '../../timestampedCollection';
-
-export const ChampionMatchups = new TimestampedCollection('ChampionMatchups');