+++ packages/app/imports/api/champions/champion-matchups/server/index.js
@@ -1,3 +1,0 @@
-import './_createIndex';
-import './_methods';
-import './_schema';