+++ packages/app/imports/api/champions/champion-stats/_collection.js
@@ -1,3 +1,0 @@
-import TimestampedCollection from '../../timestampedCollection';
-
-export const ChampionStats = new TimestampedCollection('ChampionStats');