+++ packages/app/imports/api/champions/champion-stats/server/_createIndex.js
@@ -1,9 +1,0 @@
-import { ChampionStats } from '../_collection';
-
-const rawCollection = ChampionStats.rawCollection();
-rawCollection.createIndex(
-  {
-    championId: 1
-  },
-  { unique: true, background: true }
-);