+++ packages/app/imports/api/champions/champion-stats/server/index.js
@@ -1,2 +1,0 @@
-import './_createIndex';
-import './_schema';