+++ packages/app/imports/api/champions/server/index.js
@@ -1,2 +1,0 @@
-export * from './_updateFirstBloodStats';
-export * from './_updateSnowballStats';