+++ packages/app/imports/api/matchup/helpers/_identifier.js
@@ -3,9 +3,3 @@ export const createMatchupMatchesIdentifier = ({ champ1Id, champ2Id, mapId }) =>
     ? `${champ2Id}&${champ1Id}&${mapId}`
     : `${champ1Id}&${champ2Id}&${mapId}`;
 };
-
-export const createMatchupStatsIdentifier = ({ champ1Id, champ2Id, role }) => {
-  return champ1Id > champ2Id
-    ? `${champ2Id}&${champ1Id}&${role}`
-    : `${champ1Id}&${champ2Id}&${role}`;
-};