+++ packages/app/imports/startup/app/server/register-api.js
@@ -22,8 +22,5 @@ import '/imports/api/twitch-api/server/publications';
 import '/imports/api/lotteries/server/methods';
 import '/imports/api/lotteries/server/publications';
 import '/imports/api/patreon/server/patreon';
-import '/imports/api/champions/server/methods';
 import '/imports/api/status/server/publications';
 import '/imports/api/matchup/server';
-import '/imports/api/champions/champion-matchups/server';
-import '/imports/api/champions/champion-stats/server';