+++ packages/app/imports/store/actions/_championStats.js
@@ -6,7 +6,7 @@ import {
   CLEAR_CHAMPION_STATS
 } from '../types';
 
-import { Meteor } from 'meteor/meteor';
+import axios from 'axios';
 
 export const clearChampionStats = () => {
   return {
@@ -21,11 +21,10 @@ const requestChampionStats = championId => {
   };
 };
 
-export const receiveChampionStats = (championId, { byMap, stats }) => {
+export const receiveChampionStats = (championId, stats) => {
   return {
     type: RECEIVE_CHAMPION_STATS,
     championId,
-    byMap,
     stats,
     receivedAt: Date.now()
   };
@@ -40,21 +39,153 @@ const receiveChampionStatsError = (championId, error) => {
   };
 };
 
+function sortHighestCount(a, b) {
+  return a.count - b.count;
+}
+
+function sortHighestWinrate(a, b) {
+  return a.winRate - b.winRate;
+}
+
+const roles = {
+  JUNGLE: 'JUNGLE',
+  BOTTOM: 'DUO_CARRY',
+  UTILITY: 'DUO_SUPPORT',
+  TOP: 'TOP',
+  MIDDLE: 'MIDDLE'
+};
+
+const skills = {
+  1: 'Q',
+  2: 'W',
+  3: 'E',
+  4: 'R'
+};
+
 export const fetchChampionStats = championId => {
   return dispatch => {
     dispatch(requestChampionStats(championId));
     return new Promise(resolve => {
-      Meteor.call('getChampionStats', championId, (error, result) => {
-        if (error || !result) {
-          return resolve(dispatch(receiveChampionStatsError(championId, error)));
-        }
-        return resolve(dispatch(receiveChampionStats(championId, result)));
-      });
+      axios
+        .get(`https://champs.th.gl/champs?champId=${championId}&mapId=11`)
+        .then(result => result.data)
+        .then(champion => {
+          const stats = Object.entries(champion.positions).reduce((acc, [key, stats]) => {
+            const role = roles[key];
+            const firstItems = stats.firstItems.map(obj => {
+              return {
+                count: obj.matches,
+                wins: Math.round(obj.matches * obj.winRate),
+                winRate: obj.winRate,
+                items: obj.items
+              };
+            });
+
+            const items = Object.entries(stats.items).reduce((acc, [key, items]) => {
+              return {
+                ...acc,
+                [key]: items.map(obj => ({
+                  count: obj.matches,
+                  wins: Math.round(obj.matches * obj.winRate),
+                  winRate: obj.winRate,
+                  itemId: obj.itemId
+                }))
+              };
+            }, {});
+
+            const skillOrder = stats.skillOrder.map(obj => {
+              return {
+                count: obj.matches,
+                wins: Math.round(obj.matches * obj.winRate),
+                winRate: obj.winRate,
+                order: obj.order.map(skill => skills[skill])
+              };
+            });
+
+            const runes = stats.perks.map(obj => {
+              return {
+                count: obj.matches,
+                wins: Math.round(obj.matches * obj.winRate),
+                winRate: obj.winRate,
+                order: [
+                  obj.perk0,
+                  obj.perk1,
+                  obj.perk2,
+                  obj.perk3,
+                  obj.perkSubStyle,
+                  obj.perk4,
+                  obj.perk5,
+                  obj.statPerk0,
+                  obj.statPerk1,
+                  obj.statPerk2
+                ]
+              };
+            });
+
+            return {
+              ...acc,
+              [role]: {
+                gamesPlayed: stats.matches,
+                winRate: stats.winRate,
+                playRate: champion.pickRate,
+                banRate: champion.banRate,
+                firstBloodKill: stats.averageStats.firstBloodKills,
+                snowballKills: stats.averageStats.snowballKills,
+                damageComposition: {
+                  percentTrue:
+                    stats.averageStats.trueDamageDealt / stats.averageStats.totalDamageDealt,
+                  percentMagical:
+                    stats.averageStats.magicDamageDealt / stats.averageStats.totalDamageDealt,
+                  percentPhysical:
+                    stats.averageStats.physicalDamageDealt / stats.averageStats.totalDamageDealt,
+                  totalTrue: stats.averageStats.trueDamageDealt,
+                  totalMagical: stats.averageStats.magicDamageDealt,
+                  totalPhysical: stats.averageStats.physicalDamageDealt,
+                  total: stats.averageStats.totalDamageDealt
+                },
+                firstItems: {
+                  highestCount: firstItems.sort(sortHighestCount)[0],
+                  highestWinrate: firstItems.sort(sortHighestWinrate)[0]
+                },
+                items: {
+                  highestCount: {
+                    '2-12': items['2-12'].sort(sortHighestCount).slice(0, 3),
+                    '12-22': items['12-22'].sort(sortHighestCount).slice(0, 3),
+                    '22-32': items['22-32'].sort(sortHighestCount).slice(0, 3),
+                    '32-42': items['32-42'].sort(sortHighestCount).slice(0, 3),
+                    '42-52': items['42-52'].sort(sortHighestCount).slice(0, 3)
+                  },
+                  highestWinrate: {
+                    '2-12': items['2-12'].sort(sortHighestWinrate).slice(0, 3),
+                    '12-22': items['12-22'].sort(sortHighestWinrate).slice(0, 3),
+                    '22-32': items['22-32'].sort(sortHighestWinrate).slice(0, 3),
+                    '32-42': items['32-42'].sort(sortHighestWinrate).slice(0, 3),
+                    '42-52': items['42-52'].sort(sortHighestWinrate).slice(0, 3)
+                  }
+                },
+                skillOrder: {
+                  highestCount: skillOrder.sort(sortHighestCount)[0],
+                  highestWinrate: skillOrder.sort(sortHighestWinrate)[0]
+                },
+                runes: {
+                  highestCount: runes.sort(sortHighestCount)[0],
+                  highestWinrate: runes.sort(sortHighestWinrate)[0]
+                }
+              }
+            };
+          }, {});
+
+          return resolve(dispatch(receiveChampionStats(championId, stats)));
+        })
+        .catch(error => {
+          console.error(error);
+          return resolve(dispatch(receiveChampionStatsError(championId, 'Can not fetch')));
+        });
     });
   };
 };
 
-const ttl = 1000 * 60 * 60 * 24;
+const ttl = 1000 * 60 * 3;
 export const fetchChampionStatsIfNeeded = fetchIfNeeded(
   fetchChampionStats,
   'championStatsByChampionId',