+++ packages/app/imports/ui/components/BuildOrder/BuildOrder.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
-
-import { HardwareKeyboardArrowRightIcon } from '../icons';
+import HardwareKeyboardArrowRightIcon from '@material-ui/icons/KeyboardArrowRight';
+import HardwareKeyboardArrowLeftIcon from '@material-ui/icons/KeyboardArrowLeft';
+import { IconButton } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import { Tooltip } from '@material-ui/core';
 import get from 'lodash.get';
@@ -8,6 +9,7 @@ import { getItem } from '../../../api/riot-api/staticData';
 import riotItems from '/imports/shared/riot-api/items.ts';
 import { withChampionStats } from '../../hocs/withChampionStats';
 import { TooltipTitle } from '../generic/TooltipTitle';
+import universeTheme from '../../layouts/universeTheme';
 
 const styles = {
   container: {
@@ -23,7 +25,8 @@ const styles = {
   item: {
     width: '36px',
     height: '36px',
-    verticalAlign: 'sub'
+    verticalAlign: 'sub',
+    margin: '0px 4px'
   },
   toggle: {
     position: 'absolute',
@@ -33,16 +36,39 @@ const styles = {
   },
   label: {
     margin: 2
+  },
+  leftArrow: {
+    position: 'absolute',
+    bottom: 10,
+    left: 0,
+    padding: 0,
+    height: 24,
+    width: 24,
+    color: universeTheme.palette.textColor
+  },
+  rightArrow: {
+    position: 'absolute',
+    bottom: 10,
+    right: 0,
+    padding: 0,
+    height: 24,
+    width: 24,
+    color: universeTheme.palette.textColor
   }
 };
 
+const ranges = ['2-12', '12-22', '22-32', '32-42', '42-52'];
+
 export class BuildOrder extends Component {
+  state = {
+    range: ranges[0]
+  };
+
   render() {
     const { champion, championStatsByChampionId, role, style, setValue } = this.props;
     const championStats = get(championStatsByChampionId, `${champion.id}.stats`) || {};
     const roleStats = championStats[role];
     const firstItemsStats = roleStats && roleStats.firstItems && roleStats.firstItems[setValue];
-    const finalItemsStats = roleStats && roleStats.finalItems && roleStats.finalItems[setValue];
 
     return (
       <div style={Object.assign({}, styles.container, style)}>
@@ -69,27 +95,42 @@ export class BuildOrder extends Component {
             <div style={styles.label}>Starting Build</div>
           </Tooltip>
           {this.renderFirstItems()}
+          <IconButton
+            onClick={() => {
+              this.setState({
+                range:
+                  ranges.indexOf(this.state.range) === 0
+                    ? ranges[ranges.length - 1]
+                    : ranges[ranges.indexOf(this.state.range) - 1]
+              });
+            }}
+            style={styles.leftArrow}
+          >
+            <HardwareKeyboardArrowLeftIcon />
+          </IconButton>
           <Tooltip
             title={
               <TooltipTitle
-                content={
-                  <span>
-                    Analysed matches: {finalItemsStats ? finalItemsStats.count : 0}
-                    <br />
-                    Winrate:{' '}
-                    {finalItemsStats ? `${(finalItemsStats.winRate * 100).toFixed(2)}%` : '-'}
-                  </span>
-                }
-                title={
-                  setValue === 'highestCount'
-                    ? 'Most Frequent Final Build'
-                    : 'Most Winning Final Build'
-                }
+                content={`These items are good to buy between ${this.state.range} minutes`}
+                title={setValue === 'highestCount' ? 'Most Frequent Items' : 'Most Winning Items'}
               />
             }
           >
-            <div style={styles.label}>Final Build</div>
+            <div style={styles.label}>Items between {this.state.range} minutes</div>
           </Tooltip>
+          <IconButton
+            onClick={() => {
+              this.setState({
+                range:
+                  ranges.indexOf(this.state.range) >= ranges.length - 1
+                    ? ranges[0]
+                    : ranges[ranges.indexOf(this.state.range) + 1]
+              });
+            }}
+            style={styles.rightArrow}
+          >
+            <HardwareKeyboardArrowRightIcon />
+          </IconButton>
           {this.renderFinalItems()}
         </div>
       </div>
@@ -107,19 +148,17 @@ export class BuildOrder extends Component {
     return (
       <div>
         {roleStats.firstItems[setValue].items.map((item, index) => (
-          <span key={`firstItems${index}`}>
-            {index > 0 && <HardwareKeyboardArrowRightIcon />}
-            <Tooltip
-              title={
-                <TooltipTitle
-                  content={riotItems[item] && riotItems[item].sanitizedDescription}
-                  title={riotItems[item] && riotItems[item].name}
-                />
-              }
-            >
-              <img src={getItem(`${item}.png`)} style={styles.item} />
-            </Tooltip>
-          </span>
+          <Tooltip
+            key={`firstItems${index}`}
+            title={
+              <TooltipTitle
+                content={riotItems[item] && riotItems[item].sanitizedDescription}
+                title={riotItems[item] && riotItems[item].name}
+              />
+            }
+          >
+            <img src={getItem(`${item}.png`)} style={styles.item} />
+          </Tooltip>
         ))}
       </div>
     );
@@ -129,26 +168,32 @@ export class BuildOrder extends Component {
     const { champion, championStatsByChampionId, role, setValue } = this.props;
     const championStats = get(championStatsByChampionId, `${champion.id}.stats`) || {};
     const roleStats = championStats[role];
-    if (!roleStats || !roleStats.finalItems || !roleStats.finalItems[setValue]) {
+    if (!roleStats || !roleStats.items || !roleStats.items[setValue]) {
       return <div>Not available</div>;
     }
 
     return (
       <div>
-        {roleStats.finalItems[setValue].items.map((item, index) => (
-          <span key={`items${index}`}>
-            {index > 0 && <HardwareKeyboardArrowRightIcon />}
-            <Tooltip
-              title={
-                <TooltipTitle
-                  content={riotItems[item] && riotItems[item].sanitizedDescription}
-                  title={riotItems[item] && riotItems[item].name}
-                />
-              }
-            >
-              <img src={getItem(`${item}.png`)} style={styles.item} />
-            </Tooltip>
-          </span>
+        {roleStats.items[setValue][this.state.range].map((item, index) => (
+          <Tooltip
+            key={`items${index}`}
+            title={
+              <TooltipTitle
+                content={
+                  <span>
+                    {riotItems[item.itemId] && riotItems[item.itemId].sanitizedDescription}
+                    <br />
+                    Analysed matches: {item.count}
+                    <br />
+                    Winrate: {(item.winRate * 100).toFixed(2)}%
+                  </span>
+                }
+                title={riotItems[item.itemId] && riotItems[item.itemId].name}
+              />
+            }
+          >
+            <img src={getItem(`${item.itemId}.png`)} style={styles.item} />
+          </Tooltip>
         ))}
       </div>
     );